services/dashboard/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
boto3==1.34.110
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.1
python-jose[cryptography]>=3.3.0
python-keycloak>=3.9.0
